[feat]: [CI-3052]: Add codebase variables support in  aws Vm initialize phase (#29151)
* [feat]: [CI-3052]: Add codebase variables support in aws Vm initialize phase

* fix secret value

* fix UT

* fix checkstyle

* fix tests
