+++ 320-ci-execution/src/main/java/io/harness/stateutils/buildstate/K8BuildSetupUtils.java
@@ -4,29 +4,11 @@ import static io.harness.beans.serializer.RunTimeInputHandler.UNRESOLVED_PARAMET
 import static io.harness.beans.serializer.RunTimeInputHandler.resolveIntegerParameter;
 import static io.harness.beans.serializer.RunTimeInputHandler.resolveMapParameter;
 import static io.harness.beans.serializer.RunTimeInputHandler.resolveStringParameter;
-import static io.harness.beans.sweepingoutputs.CISweepingOutputNames.CODEBASE;
 import static io.harness.beans.sweepingoutputs.CISweepingOutputNames.CODE_BASE_CONNECTOR_REF;
 import static io.harness.beans.sweepingoutputs.ContainerPortDetails.PORT_DETAILS;
 import static io.harness.beans.sweepingoutputs.PodCleanupDetails.CLEANUP_DETAILS;
 import static io.harness.beans.sweepingoutputs.StageInfraDetails.STAGE_INFRA_DETAILS;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_AWS_REGION;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_BUILD_EVENT;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR_AVATAR;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR_EMAIL;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_COMMIT_BEFORE;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_COMMIT_BRANCH;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_COMMIT_REF;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_COMMIT_SHA;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_NETRC_MACHINE;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_NETRC_PORT;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_NETRC_USERNAME;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_REMOTE_URL;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_SOURCE_BRANCH;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_TAG;
-import static io.harness.common.BuildEnvironmentConstants.DRONE_TARGET_BRANCH;
 import static io.harness.common.CIExecutionConstants.ACCOUNT_ID_ATTR;
-import static io.harness.common.CIExecutionConstants.AWS_CODE_COMMIT_URL_REGEX;
 import static io.harness.common.CIExecutionConstants.BUILD_NUMBER_ATTR;
 import static io.harness.common.CIExecutionConstants.HARNESS_ACCOUNT_ID_VARIABLE;
 import static io.harness.common.CIExecutionConstants.HARNESS_BUILD_ID_VARIABLE;
@@ -49,13 +31,7 @@ import static io.harness.common.CIExecutionConstants.PROJECT_ID_ATTR;
 import static io.harness.common.CIExecutionConstants.STAGE_ID_ATTR;
 import static io.harness.common.CIExecutionConstants.TI_SERVICE_ENDPOINT_VARIABLE;
 import static io.harness.common.CIExecutionConstants.TI_SERVICE_TOKEN_VARIABLE;
-import static io.harness.data.structure.EmptyPredicate.isEmpty;
 import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
-import static io.harness.delegate.beans.connector.ConnectorType.BITBUCKET;
-import static io.harness.delegate.beans.connector.ConnectorType.CODECOMMIT;
-import static io.harness.delegate.beans.connector.ConnectorType.GIT;
-import static io.harness.delegate.beans.connector.ConnectorType.GITHUB;
-import static io.harness.delegate.beans.connector.ConnectorType.GITLAB;
 
 import static java.lang.String.format;
 import static java.util.Collections.emptyMap;
@@ -74,7 +50,6 @@ import io.harness.beans.environment.pod.PodSetupInfo;
 import io.harness.beans.environment.pod.container.ContainerDefinitionInfo;
 import io.harness.beans.steps.stepinfo.InitializeStepInfo;
 import io.harness.beans.sweepingoutputs.CodeBaseConnectorRefSweepingOutput;
-import io.harness.beans.sweepingoutputs.CodebaseSweepingOutput;
 import io.harness.beans.sweepingoutputs.ContainerPortDetails;
 import io.harness.beans.sweepingoutputs.ContextElement;
 import io.harness.beans.sweepingoutputs.K8PodDetails;
@@ -93,31 +68,10 @@ import io.harness.delegate.beans.ci.pod.ContainerSecrets;
 import io.harness.delegate.beans.ci.pod.ImageDetailsWithConnector;
 import io.harness.delegate.beans.ci.pod.PVCParams;
 import io.harness.delegate.beans.ci.pod.SecretVariableDetails;
-import io.harness.delegate.beans.connector.ConnectorType;
-import io.harness.delegate.beans.connector.scm.GitConnectionType;
-import io.harness.delegate.beans.connector.scm.awscodecommit.AwsCodeCommitAuthType;
-import io.harness.delegate.beans.connector.scm.awscodecommit.AwsCodeCommitConnectorDTO;
-import io.harness.delegate.beans.connector.scm.awscodecommit.AwsCodeCommitHttpsAuthType;
-import io.harness.delegate.beans.connector.scm.awscodecommit.AwsCodeCommitHttpsCredentialsDTO;
-import io.harness.delegate.beans.connector.scm.awscodecommit.AwsCodeCommitUrlType;
-import io.harness.delegate.beans.connector.scm.bitbucket.BitbucketConnectorDTO;
-import io.harness.delegate.beans.connector.scm.bitbucket.BitbucketHttpAuthenticationType;
-import io.harness.delegate.beans.connector.scm.bitbucket.BitbucketHttpCredentialsDTO;
-import io.harness.delegate.beans.connector.scm.genericgitconnector.GitConfigDTO;
-import io.harness.delegate.beans.connector.scm.genericgitconnector.GitHTTPAuthenticationDTO;
-import io.harness.delegate.beans.connector.scm.github.GithubConnectorDTO;
-import io.harness.delegate.beans.connector.scm.github.GithubHttpAuthenticationType;
-import io.harness.delegate.beans.connector.scm.github.GithubHttpCredentialsDTO;
-import io.harness.delegate.beans.connector.scm.gitlab.GitlabConnectorDTO;
-import io.harness.delegate.beans.connector.scm.gitlab.GitlabHttpAuthenticationType;
-import io.harness.delegate.beans.connector.scm.gitlab.GitlabHttpCredentialsDTO;
 import io.harness.exception.ConnectorNotFoundException;
-import io.harness.exception.InvalidArgumentsException;
 import io.harness.exception.InvalidRequestException;
-import io.harness.exception.WingsException;
 import io.harness.exception.ngexception.CIStageExecutionException;
 import io.harness.ff.CIFeatureFlagService;
-import io.harness.git.GitClientHelper;
 import io.harness.k8s.model.ImageDetails;
 import io.harness.logserviceclient.CILogServiceUtils;
 import io.harness.ng.core.EntityDetail;
@@ -125,7 +79,6 @@ import io.harness.ng.core.NGAccess;
 import io.harness.pms.contracts.ambiance.Ambiance;
 import io.harness.pms.execution.utils.AmbianceUtils;
 import io.harness.pms.rbac.PipelineRbacHelper;
-import io.harness.pms.sdk.core.data.OptionalSweepingOutput;
 import io.harness.pms.sdk.core.plan.creation.yaml.StepOutcomeGroup;
 import io.harness.pms.sdk.core.resolver.RefObjectUtils;
 import io.harness.pms.sdk.core.resolver.outputs.ExecutionSweepingOutputService;
@@ -139,7 +92,6 @@ import io.harness.utils.IdentifierRefHelper;
 import io.harness.yaml.core.timeout.Timeout;
 import io.harness.yaml.extended.ci.codebase.CodeBase;
 
-import com.google.common.annotations.VisibleForTesting;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import java.time.Duration;
@@ -149,8 +101,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import net.jodah.failsafe.Failsafe;
@@ -169,6 +119,7 @@ public class K8BuildSetupUtils {
   @Inject private CIFeatureFlagService featureFlagService;
   @Inject CILogServiceUtils logServiceUtils;
   @Inject TIServiceUtils tiServiceUtils;
+  @Inject CodebaseUtils codebaseUtils;
   @Inject private PipelineRbacHelper pipelineRbacHelper;
   private final Duration RETRY_SLEEP_DURATION = Duration.ofSeconds(2);
   private final int MAX_ATTEMPTS = 3;
@@ -246,9 +197,9 @@ public class K8BuildSetupUtils {
       harnessInternalImageConnector = getDefaultInternalConnector(ngAccess);
     }
 
-    ConnectorDetails gitConnector = getGitConnector(ngAccess, ciCodebase, skipGitClone);
-    Map<String, String> gitEnvVars = getGitEnvVariables(gitConnector, ciCodebase);
-    Map<String, String> runtimeCodebaseVars = getRuntimeCodebaseVars(ambiance);
+    ConnectorDetails gitConnector = codebaseUtils.getGitConnector(ngAccess, ciCodebase, skipGitClone);
+    Map<String, String> gitEnvVars = codebaseUtils.getGitEnvVariables(gitConnector, ciCodebase);
+    Map<String, String> runtimeCodebaseVars = codebaseUtils.getRuntimeCodebaseVars(ambiance);
 
     List<CIK8ContainerParams> containerParamsList = getContainerParamsList(k8PodDetails, podSetupInfo, ngAccess,
         harnessInternalImageConnector, gitEnvVars, runtimeCodebaseVars, initializeStepInfo, logPrefix, ambiance);
@@ -312,72 +263,6 @@ public class K8BuildSetupUtils {
         .initContainerParamsList(singletonList(setupAddOnContainerParams))
         .runAsUser(stageRunAsUser)
         .build();
-  }
-
-  private Map<String, String> getRuntimeCodebaseVars(Ambiance ambiance) {
-    Map<String, String> codebaseRuntimeVars = new HashMap<>();
-
-    OptionalSweepingOutput optionalSweepingOutput =
-        executionSweepingOutputResolver.resolveOptional(ambiance, RefObjectUtils.getOutcomeRefObject(CODEBASE));
-    if (!optionalSweepingOutput.isFound()) {
-      return codebaseRuntimeVars;
-    }
-
-    CodebaseSweepingOutput codebaseSweeping = (CodebaseSweepingOutput) optionalSweepingOutput.getOutput();
-
-    /* Bitbucket SAAS does not generate refs/pull-requests/* which requires us to do this special handling.
-      Override commit ref to source branch instead of pull request ref
-     */
-    String commitRef = codebaseSweeping.getCommitRef();
-    if (isNotEmpty(codebaseSweeping.getPullRequestLink())
-        && codebaseSweeping.getPullRequestLink().contains("bitbucket.org")) {
-      commitRef = format("+refs/heads/%s", codebaseSweeping.getSourceBranch());
-    }
-
-    if (isNotEmpty(commitRef)) {
-      codebaseRuntimeVars.put(DRONE_COMMIT_REF, commitRef);
-    }
-
-    if (isNotEmpty(codebaseSweeping.getBranch())) {
-      codebaseRuntimeVars.put(DRONE_COMMIT_BRANCH, codebaseSweeping.getBranch());
-    }
-
-    if (codebaseSweeping.getBuild() != null && isNotEmpty(codebaseSweeping.getBuild().getType())
-        && codebaseSweeping.getBuild().getType().equals("PR")) {
-      codebaseRuntimeVars.put(DRONE_BUILD_EVENT, "pull_request");
-    }
-
-    if (!isEmpty(codebaseSweeping.getTag())) {
-      codebaseRuntimeVars.put(DRONE_TAG, codebaseSweeping.getTag());
-      codebaseRuntimeVars.put(DRONE_BUILD_EVENT, "tag");
-    }
-
-    if (!isEmpty(codebaseSweeping.getTargetBranch())) {
-      codebaseRuntimeVars.put(DRONE_TARGET_BRANCH, codebaseSweeping.getTargetBranch());
-    }
-
-    if (!isEmpty(codebaseSweeping.getSourceBranch())) {
-      codebaseRuntimeVars.put(DRONE_SOURCE_BRANCH, codebaseSweeping.getSourceBranch());
-    }
-
-    if (!isEmpty(codebaseSweeping.getGitUserEmail())) {
-      codebaseRuntimeVars.put(DRONE_COMMIT_AUTHOR_EMAIL, codebaseSweeping.getGitUserEmail());
-    }
-    if (!isEmpty(codebaseSweeping.getGitUserAvatar())) {
-      codebaseRuntimeVars.put(DRONE_COMMIT_AUTHOR_AVATAR, codebaseSweeping.getGitUserAvatar());
-    }
-
-    if (!isEmpty(codebaseSweeping.getGitUserId())) {
-      codebaseRuntimeVars.put(DRONE_COMMIT_AUTHOR, codebaseSweeping.getGitUserId());
-    }
-    if (isNotEmpty(codebaseSweeping.getBaseCommitSha())) {
-      codebaseRuntimeVars.put(DRONE_COMMIT_BEFORE, codebaseSweeping.getBaseCommitSha());
-    }
-    if (isNotEmpty(codebaseSweeping.getCommitSha())) {
-      codebaseRuntimeVars.put(DRONE_COMMIT_SHA, codebaseSweeping.getCommitSha());
-    }
-
-    return codebaseRuntimeVars;
   }
 
   public List<CIK8ContainerParams> getContainerParamsList(K8PodDetails k8PodDetails, PodSetupInfo podSetupInfo,
@@ -672,195 +557,6 @@ public class K8BuildSetupUtils {
     envVars.put(HARNESS_STAGE_ID_VARIABLE, stageID);
     envVars.put(HARNESS_LOG_PREFIX_VARIABLE, logPrefix);
     return envVars;
-  }
-
-  @VisibleForTesting
-  Map<String, String> getGitEnvVariables(ConnectorDetails gitConnector, CodeBase ciCodebase) {
-    Map<String, String> envVars = new HashMap<>();
-    if (gitConnector == null) {
-      return envVars;
-    }
-
-    validateGitConnector(gitConnector);
-    if (gitConnector.getConnectorType() == GITHUB) {
-      GithubConnectorDTO gitConfigDTO = (GithubConnectorDTO) gitConnector.getConnectorConfig();
-      validateGithubConnectorAuth(gitConfigDTO);
-      envVars = retrieveGitSCMEnvVar(ciCodebase, gitConfigDTO.getConnectionType(), gitConfigDTO.getUrl());
-    } else if (gitConnector.getConnectorType() == GITLAB) {
-      GitlabConnectorDTO gitConfigDTO = (GitlabConnectorDTO) gitConnector.getConnectorConfig();
-      validateGitlabConnectorAuth(gitConfigDTO);
-      envVars = retrieveGitSCMEnvVar(ciCodebase, gitConfigDTO.getConnectionType(), gitConfigDTO.getUrl());
-    } else if (gitConnector.getConnectorType() == BITBUCKET) {
-      BitbucketConnectorDTO gitConfigDTO = (BitbucketConnectorDTO) gitConnector.getConnectorConfig();
-      validateBitbucketConnectorAuth(gitConfigDTO);
-      envVars = retrieveGitSCMEnvVar(ciCodebase, gitConfigDTO.getConnectionType(), gitConfigDTO.getUrl());
-    } else if (gitConnector.getConnectorType() == CODECOMMIT) {
-      AwsCodeCommitConnectorDTO gitConfigDTO = (AwsCodeCommitConnectorDTO) gitConnector.getConnectorConfig();
-      envVars = retrieveAwsCodeCommitEnvVar(gitConfigDTO, ciCodebase);
-    } else if (gitConnector.getConnectorType() == GIT) {
-      GitConfigDTO gitConfigDTO = (GitConfigDTO) gitConnector.getConnectorConfig();
-      envVars = retrieveGitEnvVar(gitConfigDTO, ciCodebase);
-    } else {
-      throw new CIStageExecutionException("Unsupported git connector type" + gitConnector.getConnectorType());
-    }
-
-    return envVars;
-  }
-
-  private Map<String, String> retrieveGitSCMEnvVar(CodeBase ciCodebase, GitConnectionType connectionType, String url) {
-    Map<String, String> envVars = new HashMap<>();
-    String gitUrl = IntegrationStageUtils.getGitURL(ciCodebase, connectionType, url);
-    String domain = GitClientHelper.getGitSCM(gitUrl);
-    String port = GitClientHelper.getGitSCMPort(gitUrl);
-    if (port != null) {
-      envVars.put(DRONE_NETRC_PORT, port);
-    }
-
-    envVars.put(DRONE_REMOTE_URL, gitUrl);
-    envVars.put(DRONE_NETRC_MACHINE, domain);
-    return envVars;
-  }
-
-  private void validateGithubConnectorAuth(GithubConnectorDTO gitConfigDTO) {
-    switch (gitConfigDTO.getAuthentication().getAuthType()) {
-      case HTTP:
-        GithubHttpCredentialsDTO gitAuth = (GithubHttpCredentialsDTO) gitConfigDTO.getAuthentication().getCredentials();
-        if (gitAuth.getType() != GithubHttpAuthenticationType.USERNAME_AND_PASSWORD
-            && gitAuth.getType() != GithubHttpAuthenticationType.USERNAME_AND_TOKEN) {
-          throw new CIStageExecutionException("Unsupported github connector auth type" + gitAuth.getType());
-        }
-        break;
-      case SSH:
-        break;
-      default:
-        throw new CIStageExecutionException(
-            "Unsupported github connector auth" + gitConfigDTO.getAuthentication().getAuthType());
-    }
-  }
-
-  private void validateBitbucketConnectorAuth(BitbucketConnectorDTO gitConfigDTO) {
-    switch (gitConfigDTO.getAuthentication().getAuthType()) {
-      case HTTP:
-        BitbucketHttpCredentialsDTO gitAuth =
-            (BitbucketHttpCredentialsDTO) gitConfigDTO.getAuthentication().getCredentials();
-        if (gitAuth.getType() != BitbucketHttpAuthenticationType.USERNAME_AND_PASSWORD) {
-          throw new CIStageExecutionException("Unsupported bitbucket connector auth type" + gitAuth.getType());
-        }
-        break;
-      case SSH:
-        break;
-      default:
-        throw new CIStageExecutionException(
-            "Unsupported bitbucket connector auth" + gitConfigDTO.getAuthentication().getAuthType());
-    }
-  }
-
-  private void validateGitlabConnectorAuth(GitlabConnectorDTO gitConfigDTO) {
-    switch (gitConfigDTO.getAuthentication().getAuthType()) {
-      case HTTP:
-        GitlabHttpCredentialsDTO gitAuth = (GitlabHttpCredentialsDTO) gitConfigDTO.getAuthentication().getCredentials();
-        if (gitAuth.getType() != GitlabHttpAuthenticationType.USERNAME_AND_PASSWORD
-            && gitAuth.getType() != GitlabHttpAuthenticationType.USERNAME_AND_TOKEN) {
-          throw new CIStageExecutionException("Unsupported gitlab connector auth type" + gitAuth.getType());
-        }
-        break;
-      case SSH:
-        break;
-      default:
-        throw new CIStageExecutionException(
-            "Unsupported gitlab connector auth" + gitConfigDTO.getAuthentication().getAuthType());
-    }
-  }
-
-  private Map<String, String> retrieveGitEnvVar(GitConfigDTO gitConfigDTO, CodeBase ciCodebase) {
-    Map<String, String> envVars = new HashMap<>();
-    String gitUrl =
-        IntegrationStageUtils.getGitURL(ciCodebase, gitConfigDTO.getGitConnectionType(), gitConfigDTO.getUrl());
-    String domain = GitClientHelper.getGitSCM(gitUrl);
-
-    envVars.put(DRONE_REMOTE_URL, gitUrl);
-    envVars.put(DRONE_NETRC_MACHINE, domain);
-    switch (gitConfigDTO.getGitAuthType()) {
-      case HTTP:
-        GitHTTPAuthenticationDTO gitAuth = (GitHTTPAuthenticationDTO) gitConfigDTO.getGitAuth();
-        envVars.put(DRONE_NETRC_USERNAME, gitAuth.getUsername());
-        break;
-      case SSH:
-        break;
-      default:
-        throw new CIStageExecutionException("Unsupported bitbucket connector auth" + gitConfigDTO.getGitAuthType());
-    }
-    return envVars;
-  }
-
-  private Map<String, String> retrieveAwsCodeCommitEnvVar(AwsCodeCommitConnectorDTO gitConfigDTO, CodeBase ciCodebase) {
-    Map<String, String> envVars = new HashMap<>();
-    GitConnectionType gitConnectionType =
-        gitConfigDTO.getUrlType() == AwsCodeCommitUrlType.REPO ? GitConnectionType.REPO : GitConnectionType.ACCOUNT;
-    String gitUrl = IntegrationStageUtils.getGitURL(ciCodebase, gitConnectionType, gitConfigDTO.getUrl());
-
-    envVars.put(DRONE_REMOTE_URL, gitUrl);
-    envVars.put(DRONE_AWS_REGION, getAwsCodeCommitRegion(gitConfigDTO.getUrl()));
-    if (gitConfigDTO.getAuthentication().getAuthType() == AwsCodeCommitAuthType.HTTPS) {
-      AwsCodeCommitHttpsCredentialsDTO credentials =
-          (AwsCodeCommitHttpsCredentialsDTO) gitConfigDTO.getAuthentication().getCredentials();
-      if (credentials.getType() != AwsCodeCommitHttpsAuthType.ACCESS_KEY_AND_SECRET_KEY) {
-        throw new CIStageExecutionException("Unsupported aws code commit connector auth type" + credentials.getType());
-      }
-    } else {
-      throw new CIStageExecutionException(
-          "Unsupported aws code commit connector auth" + gitConfigDTO.getAuthentication().getAuthType());
-    }
-    return envVars;
-  }
-
-  private String getAwsCodeCommitRegion(String url) {
-    Pattern r = Pattern.compile(AWS_CODE_COMMIT_URL_REGEX);
-    Matcher m = r.matcher(url);
-
-    if (m.find()) {
-      return m.group(1);
-    } else {
-      throw new InvalidRequestException("Url does not have region information");
-    }
-  }
-
-  private void validateGitConnector(ConnectorDetails gitConnector) {
-    if (gitConnector == null) {
-      log.error("Git connector is not valid {}", gitConnector);
-      throw new InvalidArgumentsException("Git connector is not valid", WingsException.USER);
-    }
-    if (gitConnector.getConnectorType() != GIT && gitConnector.getConnectorType() != ConnectorType.GITHUB
-        && gitConnector.getConnectorType() != ConnectorType.GITLAB && gitConnector.getConnectorType() != BITBUCKET
-        && gitConnector.getConnectorType() != CODECOMMIT) {
-      log.error("Git connector ref is not of type git {}", gitConnector.getConnectorType());
-      throw new InvalidArgumentsException(
-          "Connector type is not from supported connectors list GITHUB, GITLAB, BITBUCKET, CODECOMMIT ",
-          WingsException.USER);
-    }
-
-    // TODO Validate all
-
-    //    GitConfigDTO gitConfigDTO = (GitConfigDTO) gitConnector.getConnectorConfig();
-    //    if (gitConfigDTO.getGitAuthType() != GitAuthType.HTTP && gitConfigDTO.getGitAuthType() != GitAuthType.SSH) {
-    //      log.error("Git connector ref is of invalid auth type {}", gitConnector);
-    //      throw new InvalidArgumentsException("Invalid auth provided for git connector", WingsException.USER);
-    //    }
-  }
-
-  private ConnectorDetails getGitConnector(NGAccess ngAccess, CodeBase codeBase, boolean skipGitClone) {
-    if (skipGitClone) {
-      return null;
-    }
-
-    if (codeBase == null) {
-      throw new CIStageExecutionException("CI codebase is mandatory in case git clone is enabled");
-    }
-
-    if (codeBase.getConnectorRef() == null) {
-      throw new CIStageExecutionException("Git connector is mandatory in case git clone is enabled");
-    }
-    return connectorUtils.getConnectorDetails(ngAccess, codeBase.getConnectorRef());
   }
 
   private Map<String, String> getBuildLabels(Ambiance ambiance, K8PodDetails k8PodDetails) {
