+++ 930-delegate-tasks/src/main/java/io/harness/delegate/task/citasks/vm/CIVmInitializeTaskHandler.java
@@ -1,5 +1,6 @@
 package io.harness.delegate.task.citasks.vm;
 
+import static io.harness.data.encoding.EncodingUtils.decodeBase64;
 import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
 import static io.harness.delegate.task.citasks.vm.helper.CIVMConstants.DRONE_COMMIT_BRANCH;
 import static io.harness.delegate.task.citasks.vm.helper.CIVMConstants.DRONE_COMMIT_LINK;
@@ -16,18 +17,22 @@ import static java.lang.String.format;
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.delegate.beans.ci.CIInitializeTaskParams;
+import io.harness.delegate.beans.ci.pod.SecretParams;
 import io.harness.delegate.beans.ci.vm.CIVmInitializeTaskParams;
 import io.harness.delegate.beans.ci.vm.VmTaskExecutionResponse;
 import io.harness.delegate.beans.ci.vm.runner.SetupVmRequest;
 import io.harness.delegate.beans.ci.vm.runner.SetupVmResponse;
 import io.harness.delegate.beans.logstreaming.ILogStreamingTaskClient;
 import io.harness.delegate.task.citasks.CIInitializeTaskHandler;
+import io.harness.delegate.task.citasks.cik8handler.SecretSpecBuilder;
 import io.harness.delegate.task.citasks.vm.helper.HttpHelper;
 import io.harness.logging.CommandExecutionStatus;
 
 import com.google.inject.Inject;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import javax.validation.constraints.NotNull;
 import lombok.extern.slf4j.Slf4j;
@@ -38,6 +43,7 @@ import retrofit2.Response;
 public class CIVmInitializeTaskHandler implements CIInitializeTaskHandler {
   @NotNull private Type type = CIInitializeTaskHandler.Type.VM;
   @Inject private HttpHelper httpHelper;
+  @Inject private SecretSpecBuilder secretSpecBuilder;
 
   @Override
   public Type getType() {
@@ -46,10 +52,10 @@ public class CIVmInitializeTaskHandler implements CIInitializeTaskHandler {
 
   public VmTaskExecutionResponse executeTaskInternal(
       CIInitializeTaskParams ciInitializeTaskParams, ILogStreamingTaskClient logStreamingTaskClient) {
-    CIVmInitializeTaskParams CIVmInitializeTaskParams = (CIVmInitializeTaskParams) ciInitializeTaskParams;
+    CIVmInitializeTaskParams ciVmInitializeTaskParams = (CIVmInitializeTaskParams) ciInitializeTaskParams;
     log.info(
-        "Received request to initialize stage with stage runtime ID {}", CIVmInitializeTaskParams.getStageRuntimeId());
-    return callRunnerForSetup(CIVmInitializeTaskParams);
+        "Received request to initialize stage with stage runtime ID {}", ciVmInitializeTaskParams.getStageRuntimeId());
+    return callRunnerForSetup(ciVmInitializeTaskParams);
   }
 
   private VmTaskExecutionResponse callRunnerForSetup(CIVmInitializeTaskParams ciVmInitializeTaskParams) {
@@ -78,10 +84,19 @@ public class CIVmInitializeTaskHandler implements CIInitializeTaskHandler {
 
   private SetupVmRequest convert(CIVmInitializeTaskParams params) {
     Map<String, String> env = new HashMap<>();
+    List<String> secrets = new ArrayList<>();
     if (isNotEmpty(params.getEnvironment())) {
       env = params.getEnvironment();
     }
 
+    if (params.getGitConnector() != null) {
+      Map<String, SecretParams> secretVars = secretSpecBuilder.decryptGitSecretVariables(params.getGitConnector());
+      for (Map.Entry<String, SecretParams> entry : secretVars.entrySet()) {
+        String secret = new String(decodeBase64(entry.getValue().getValue()));
+        env.put(entry.getKey(), secret);
+        secrets.add(secret);
+      }
+    }
     SetupVmRequest.TIConfig tiConfig = SetupVmRequest.TIConfig.builder()
                                            .url(params.getTiUrl())
                                            .token(params.getTiSvcToken())
@@ -108,7 +123,8 @@ public class CIVmInitializeTaskHandler implements CIInitializeTaskHandler {
                                            .build();
 
     SetupVmRequest.Config config = SetupVmRequest.Config.builder()
-                                       .envs(params.getEnvironment())
+                                       .envs(env)
+                                       .secrets(secrets)
                                        .network(SetupVmRequest.Network.builder().id(NETWORK_ID).build())
                                        .logConfig(SetupVmRequest.LogConfig.builder()
                                                       .url(params.getLogStreamUrl())
